ics_generator-master/requirements.txt
autopep8==1.5.7
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
certifi==2021.5.30
charset-normalizer==2.0.4
icalendar==4.0.7
idna==3.2
O365==2.0.15
oauthlib==3.1.1
pycodestyle==2.7.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.2.1
stringcase==1.2.0
toml==0.10.2
tzlocal==3.0
urllib3==1.26.6
